chore: create a process of the PS/2 keyboard
