+++ kernel/Cargo.toml
@@ -46,3 +46,4 @@ page_box = { path = "../page_box" }
 terminal = { path = "../terminal" }
 xhci = "0.5.0"
 accessor = "0.3.0"
+ps2_keyboard = { path = "../ps2_keyboard" }
