+++ kernel/src/device/mod.rs
@@ -1,5 +1,4 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-pub mod keyboard;
 pub mod mouse;
 pub mod pci;
