+++ kernel/src/lib.rs
@@ -41,7 +41,7 @@ mod tss;
 
 use common::{constant::INITRD_ADDR, kernelboot};
 use device::{
-    keyboard, mouse,
+    mouse,
     pci::{ahci, xhci},
 };
 use fs::ustar::Ustar;
@@ -104,6 +104,7 @@ fn initialize_in_user_mode(boot_info: &mut kernelboot::Info) {
     process::manager::init();
 
     process::manager::add(run_tasks, Privilege::User);
+    process::manager::add(ps2_keyboard::main, Privilege::User);
     process::manager::add(tsukemen::main, Privilege::User);
 
     if cfg!(feature = "qemu_test") {
@@ -124,9 +125,8 @@ fn wait_until_timer_interrupt_happens() -> ! {
 }
 
 fn run_tasks() {
-    multitask::add(Task::new(keyboard::task()));
     multitask::add(Task::new(mouse::task()));
-    multitask::add(Task::new(xhci::task()));
+    // multitask::add(Task::new(xhci::task()));
     multitask::add(Task::new(ahci::task()));
 
     let mut executor = Executor::new();
