+++ ps2_keyboard/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ps2_keyboard"
+version = "0.1.0"
+authors = ["toku-sa-n <tokusan441@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+common = { path = "../common" }
+syscalls = { path = "../syscalls" }
